layers: Add shader_attribute_validation.cpp
